cleaning up main
moved functions to functions.cpp
adding comments
