+++ README.md
@@ -1,1 +1,15 @@
-# test
+----------------------
+Command Line Arguments
+----------------------
+Example:  ./scheduler filename
+
+
+------------------------
+Design Decision & Issues
+------------------------
+
+
+----------------------
+Performance Analysis
+----------------------
+
