+++ test1.json
@@ -1,0 +1,26 @@
+[
+    {
+        "name": "badassm<>o/t'sher",
+        "prerequisites": ["Differential Equations", "Intro to Physics","Scientific Thinking","Calculus"]
+    },
+    {
+        "name": "Calculus",
+        "prerequisites": []
+    },
+    {
+        "name": "Scientific Thinking",
+        "prerequisites": []
+    },
+    {
+        "name": "Differential Equations",
+        "prerequisites": ["Calculus"]
+    },
+    {
+        "name": "Intro to Physics",
+        "prerequisites": ["Scientific Thinking"]
+    },
+    {
+        "name": "Relativity",
+        "prerequisites": ["Differential Equations", "Intro to Physics"]
+    }
+]
